Moving the print functions to a new package: printer
